+++ packages/keyring-api/CHANGELOG.md
@@ -10,8 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add support for `base32` private key encoding in `exportAccount` ([#589](https://github.com/MetaMask/accounts/pull/589))
-- Add `Balance.metadata` ([#601](https://github.com/MetaMask/accounts/pull/601))
+- Add `{Balance,FunibleAssetAmountStruct}.metadata` ([#601](https://github.com/MetaMask/accounts/pull/601))
   - This can be used to pass extra-metadata alongside assets balances.
+  - `FungibleAssetAmountStruct` now composes `BalanceStruct`
+  - `notify:accountBalancesUpdated` also accepts `metadata` now.
 
 ### Changed